hard 문제 풀이, normal 2번 수정

+++ src/main/java/com/ohgiraffers/section01/conditional/level02/normal/Application2.java
@@ -20,9 +20,9 @@ public class Application2 {
         Scanner sc = new Scanner(System.in);
 
         System.out.println("키를 입력해주세요(cm) : ");
-        double height = sc.nextInt();
+        double height = sc.nextDouble();
         System.out.println("몸무게를 입력해주세요(kg) : ");
-        double weight = sc.nextInt();
+        double weight = sc.nextDouble();
 
         double bmi = weight / ((height / 100) * (height / 100));
